+++ inst/sql/range_join.sql
@@ -3,9 +3,9 @@ DROP TABLE IF EXISTS {{to}};
 CREATE TEMP TABLE {{to}} AS
 SELECT
     {{key}}.*,
-    {{join}}.* EXCLUDE ('#CHROM')
+    {{join}}.* EXCLUDE (CHROM)
 FROM {{key}}
 JOIN {{join}}
-    ON {{join}}."#CHROM" = {{key}}."#CHROM"
+    ON {{join}}.CHROM = {{key}}.CHROM
     AND {{join}}.start <= {{key}}.POS
     AND {{join}}.end >= {{key}}.POS;