+++ tests/testthat/test-GenomicRanges.R
@@ -8,7 +8,7 @@ test_that("to_GPos() works", {
 
     chr <- paste0("chr", rep(1:2, c(3, 5)))
     tbl <- tibble(
-        `#CHROM` = chr,
+        CHROM = chr,
         POS = 1:8,
         extra = letters[1:8],
         genome = "hg38"
@@ -22,14 +22,14 @@ test_that("to_GPos() works", {
     expect_identical(gpos$extra, letters[1:8])
     expect_identical(names(genome(gpos)), paste0("chr", c(1:22, "X", "Y", "M")))
 
-    tbl <- tibble(`#CHROM` = character(), POS = integer(), genome = character())
+    tbl <- tibble(CHROM = character(), POS = integer(), genome = character())
     expect_no_condition(gpos <- to_GPos(tbl))
     expect_identical(NROW(gpos), 0L)
     expect_identical(genome(gpos), stats::setNames(nm = character()))
     expect_identical(colnames(mcols(gpos)), character())
 
     tbl <- tibble(
-        `#CHROM` = character(), POS = integer(), genome = character(),
+        CHROM = character(), POS = integer(), genome = character(),
         extra = character()
     )
     expect_no_condition(gpos <- to_GPos(tbl))
@@ -37,12 +37,12 @@ test_that("to_GPos() works", {
     expect_identical(mcols(gpos)$extra, character(0))
 
     tbl <- tibble(
-        `CHROM` = character(), POS = integer(), genome = character()
+        `#CHROM` = character(), POS = integer(), genome = character()
     )
     expect_error(to_GPos(tbl))
 
     tbl <- tibble(
-        `#CHROM` = character(), pos = integer(), genome = character()
+        CHROM = character(), pos = integer(), genome = character()
     )
     expect_error(to_GPos(tbl))
     tbl <- tibble()