+++ NEWS.md
@@ -1,3 +1,5 @@
 # AlphaMissense 0.99.0
 
-* Initial Bioconductor submission.
+* (v. 0.0.18) Rename (including updating existing tables) '#CHROM' to
+  'CHROM' in hg19 / hg38 tables
+* (v. 0.0.17) Initial Bioconductor submission.